+++ tests/test_api.py
@@ -643,3 +643,44 @@ def test_basic_unknown_owner_1():
             assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 1
         else:
             assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 0
+
+
+def test_basic_duplicate_statement_id_1():
+
+    cove_temp_folder = tempfile.mkdtemp(prefix='lib-cove-bods-tests-', dir=tempfile.gettempdir())
+    json_filename = os.path.join(os.path.dirname(
+        os.path.realpath(__file__)), 'fixtures', 'api', 'basic_duplicate_statement_id_1.json'
+    )
+
+    results = bods_json_output(cove_temp_folder, json_filename)
+
+    assert results['validation_errors_count'] == 0
+    assert results['additional_fields_count'] == 0
+    assert results['additional_checks_count'] == 1
+
+    assert results['file_type'] == 'json'
+    assert results['statistics']['count_entity_statements'] == 1
+    for k in results['statistics']['count_entity_statements_types']:
+        if k == 'registeredEntity':
+            assert results['statistics']['count_entity_statements_types'][k] == 1
+        else:
+            assert results['statistics']['count_entity_statements_types'][k] == 0
+    assert results['statistics']['count_person_statements'] == 1
+    for k in results['statistics']['count_person_statements_types']:
+        if k == 'knownPerson':
+            assert results['statistics']['count_person_statements_types'][k] == 1
+        else:
+            assert results['statistics']['count_person_statements_types'][k] == 0
+    assert results['statistics']['count_ownership_or_control_statement'] == 1
+    assert results['statistics']['count_ownership_or_control_statement_interested_party_with_person'] == 1
+    assert results['statistics']['count_ownership_or_control_statement_interested_party_with_entity'] == 0
+    assert results['statistics']['count_ownership_or_control_statement_interested_party_with_unspecified'] == 0
+    for k in results['statistics']['count_ownership_or_control_statement_interest_statement_types']:
+        if k == 'shareholding':
+            assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 1
+        else:
+            assert results['statistics']['count_ownership_or_control_statement_interest_statement_types'][k] == 0
+
+    assert results['additional_checks'][0]['type'] == 'duplicate_statement_id'
+    assert results['additional_checks'][0]['id'] == '019a93f1-e470-42e9-957b-03559861b2e2'
+    assert results['additional_checks'][0]['count'] == 2